requirements.txt
torch==2.5.1
torchvision==0.20.1
tensorflow-macos==2.16.1
transformers==4.36.0
datasets==2.16.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
langchain==0.1.0
langchain-community==0.0.10
chromadb==0.4.22
sentence-transformers==2.2.2
openai==1.6.1
fastapi==0.108.0
uvicorn==0.25.0
pydantic==2.5.3
python-multipart==0.0.6
mlflow==2.9.2
prometheus-client==0.19.0
python-dotenv==1.0.0
requests==2.31.0
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0